+++ ProdManager/models/Incident.py
@@ -152,3 +152,6 @@ class Incident(db.Model):
       result[key] = value
 
     return result
+
+  def notify_attachments(self):
+    return []